requirements.txt
Django==2.1.2
django-rest-auth==0.9.3
djangorestframework==3.8.2
profanity==1.1
pytz==2018.5
six==1.11.0
PyJWT
ujson
djangorestframework-jwt
django-cors-headers